Requirements Complete
Game checks for wins from either player, and calls cats-game if the board is full and neither has four in a row.
players can chat with one another.
if a player leaves, the server will seek through the list of connected clients and try to find one that isn't currently playing to fill the slot.
when a player wins, the client will show a game over panel where either player can call for a rematch using the new 'RMCH' packet, resetting the board. The server sends an 'UPDT' packet with fresh board returning the users to the gameplay panel with a new game.